uses courses from the database
Now the custom study-plan uses the courses from the dummy database, which can easely  be configured when the proper database is set up.

The correct info is shown and the courses can be dragged.

The "available courses" has been styled to allways be on the right, no matter how many courses are present, and is scrollable.

There is an issue when dragging courses from "avaialable courses" list, which makes them disappear outside that list.